requirements.txt
pandas
numpy
matplotlib
scikit-learn
graphviz
pycaret==3.3.0
category_encoders
modelbit
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.4
PyYAML
httpx
-e .
